+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 - Fix remote player camera detaching after teleporting to the same map
 - Fix remote external animations (like lea sitting)
+- Fix crash on ig.EVENT_STEP.ADD_GUI
 
 ## [0.7.2] 2025-12-19