+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Fix crash on ig.EVENT_STEP.ADD_GUI
 - Fix pvp hp bars not disappearing after player defeat on remote
 - Fix pvp hp bars not rearranging themselves after player defeat on remote
+- Fix remote npc animations sometimes not being correct
 
 ## [0.7.2] 2025-12-19